requirements.txt
absl-py==0.9.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
fire==0.2.1
google-auth==1.11.0
google-auth-oauthlib==0.4.1
grpcio==1.26.0
idna==2.8
imageio==2.6.1
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
oauthlib==3.1.0
omegaconf==1.4.1
opencv-python==4.2.0.32
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
PyQt5==5.13.2
PyQt5-sip==4.19.20
python-dateutil==2.8.1
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
tensorboard==2.1.0
tensorboardX==2.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
urllib3==1.25.8
Werkzeug==0.16.1
xgboost==0.90
